[#527] change package name backend (#530)
