--- android/src/huawei/kotlin/com/github/mustafaozhan/ccc/android/util/AppReviewUtil.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2020 Mustafa Ozhan. All rights reserved.
- */
-package com.github.mustafaozhan.ccc.android.util
-
-import android.app.Activity
-
-@Suppress("UNUSED_PARAMETER")
-fun requestAppReview(activity: Activity) = Unit
