+++ android/src/huawei/kotlin/com/oztechan/ccc/android/util/AppReviewUtil.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2020 Mustafa Ozhan. All rights reserved.
+ */
+package com.oztechan.ccc.android.util
+
+import android.app.Activity
+
+@Suppress("UNUSED_PARAMETER")
+fun requestAppReview(activity: Activity) = Unit
