+++ android/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="com.github.mustafaozhan.ccc.android.app.CCCApplication"
+        android:name="com.oztechan.ccc.android.app.CCCApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -20,7 +20,7 @@
         tools:targetApi="m">
 
         <activity
-            android:name="com.github.mustafaozhan.ccc.android.ui.main.MainActivity"
+            android:name="com.oztechan.ccc.android.ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
