+++ android/src/main/kotlin/com/oztechan/ccc/android/ui/currencies/CurrenciesAdapter.kt
@@ -1,0 +1,64 @@
+package com.oztechan.ccc.android.ui.currencies
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import android.view.animation.AnimationUtils
+import androidx.recyclerview.widget.DiffUtil
+import com.github.mustafaozhan.basemob.adapter.BaseVBRecyclerViewAdapter
+import com.github.mustafaozhan.ccc.client.model.Currency
+import com.github.mustafaozhan.ccc.client.viewmodel.currencies.CurrenciesEvent
+import com.oztechan.ccc.android.util.setBackgroundByName
+import mustafaozhan.github.com.mycurrencies.R
+import mustafaozhan.github.com.mycurrencies.databinding.ItemCurrenciesBinding
+
+class CurrenciesAdapter(
+    private val currenciesEvent: CurrenciesEvent
+) : BaseVBRecyclerViewAdapter<Currency>(CurrenciesDiffer()) {
+
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int
+    ) = RatesVBViewHolder(
+        ItemCurrenciesBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            false
+        )
+    )
+
+    override fun onBindViewHolder(holder: BaseVBViewHolder<Currency>, position: Int) {
+        holder.itemView.startAnimation(
+            AnimationUtils.loadAnimation(
+                holder.itemView.context,
+                R.anim.fall_down
+            )
+        )
+        super.onBindViewHolder(holder, position)
+    }
+
+    override fun onViewDetachedFromWindow(holder: BaseVBViewHolder<Currency>) {
+        super.onViewDetachedFromWindow(holder)
+        holder.itemView.clearAnimation()
+    }
+
+    inner class RatesVBViewHolder(private val itemBinding: ItemCurrenciesBinding) :
+        BaseVBViewHolder<Currency>(itemBinding) {
+
+        override fun onItemBind(item: Currency) = with(itemBinding) {
+            imgIcon.setBackgroundByName(item.name)
+            txtSettingItem.text = item.getVariablesOneLine()
+            checkBox.isChecked = item.isActive
+            root.setOnClickListener { currenciesEvent.onItemClick(item) }
+            root.setOnLongClickListener {
+                currenciesEvent.onItemLongClick()
+                true
+            }
+        }
+    }
+
+    class CurrenciesDiffer : DiffUtil.ItemCallback<Currency>() {
+        override fun areItemsTheSame(oldItem: Currency, newItem: Currency) = true
+
+        override fun areContentsTheSame(oldItem: Currency, newItem: Currency) = false
+    }
+}
