+++ android/src/main/res/navigation/main_graph.xml
@@ -9,7 +9,7 @@
 
     <fragment
         android:id="@+id/calculatorFragment"
-        android:name="com.github.mustafaozhan.ccc.android.ui.calculator.CalculatorFragment"
+        android:name="com.oztechan.ccc.android.ui.calculator.CalculatorFragment"
         android:label="CalculatorFragment"
         tools:layout="@layout/fragment_calculator">
         <action
@@ -25,7 +25,7 @@
 
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="com.github.mustafaozhan.ccc.android.ui.settings.SettingsFragment"
+        android:name="com.oztechan.ccc.android.ui.settings.SettingsFragment"
         android:label="SettingsFragment"
         tools:layout="@layout/fragment_settings">
         <action
@@ -39,13 +39,13 @@
 
     <dialog
         android:id="@+id/adRemoveBottomSheet"
-        android:name="com.github.mustafaozhan.ccc.android.ui.adremove.AdRemoveBottomSheet"
+        android:name="com.oztechan.ccc.android.ui.adremove.AdRemoveBottomSheet"
         android:label="AdRemoveBottomSheet"
         tools:layout="@layout/bottom_sheet_ad_remove" />
 
     <fragment
         android:id="@+id/currenciesFragment"
-        android:name="com.github.mustafaozhan.ccc.android.ui.currencies.CurrenciesFragment"
+        android:name="com.oztechan.ccc.android.ui.currencies.CurrenciesFragment"
         android:label="CurrenciesFragment"
         tools:layout="@layout/fragment_currencies">
 
@@ -59,7 +59,7 @@
 
     <dialog
         android:id="@+id/changeBaseBottomSheet"
-        android:name="com.github.mustafaozhan.ccc.android.ui.changebase.ChangeBaseBottomSheet"
+        android:name="com.oztechan.ccc.android.ui.changebase.ChangeBaseBottomSheet"
         android:label="ChangeBaseBottomSheet"
         tools:layout="@layout/bottom_sheet_change_base">
         <action
@@ -69,7 +69,7 @@
 
     <fragment
         android:id="@+id/sliderFragment"
-        android:name="com.github.mustafaozhan.ccc.android.ui.slider.SliderFragment"
+        android:name="com.oztechan.ccc.android.ui.slider.SliderFragment"
         android:label="SliderFragment"
         tools:layout="@layout/fragment_slider">
